Rename TestedCasesTable to TestedCasesDetailsTable
